Apply 400_Temp.patch

+++ src/settings_type.h
@@ -283,7 +283,7 @@ struct RainfallSettings {
 	int16 wider_valleys_randomness;                       ///< Randomness when deciding on the width of wider valleys.
 
 	uint32 small_oceans_removal_factor;                    ///< Small oceans will be removed by terraforming, if they are smaller than MapSize() divided by this value.  Set to zero to turn this feature off.
-	uint16 small_basins_removal_limit;                     ///< Small basins will be removed by terraforming, if they are smaller than this number of tiles.
+	uint32 small_basins_removal_limit;                     ///< Small basins will be removed by terraforming, if they are smaller than this number of tiles.
 
 	uint16 lake_outflow_canyon_probability;                ///< probability (measured in 1/1000) that an outflow canyon is digged for the lake at hand
 	uint16 lake_reduce_to_guaranteed_probability;          ///< probability (measured in 1/1000) that a lake is reduced to its guaranteed tiles
